requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
fake-useragent==1.4.0
lxml==4.9.3
webdriver-manager==4.0.1
playwright==1.40.0
aiohttp==3.9.0
python-linkedin-v2==2.0.1a0
social-auth-app-django==5.4.0
django-extensions==3.2.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==24.2
openpyxl==3.1.5
